Began UI work and started editing of backend of serial interface to use enums.

+++ src/Serial/TXCommand.java
@@ -1,7 +1,7 @@
 package Serial;
 
 public interface TXCommand extends Command {//Interface for commands GOING TO Arduino.
-    int getCommandByte();
+    byte getCommandByte();
 
     byte[] getByteArray();
